kusal-pdf-converter/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pdf2docx==0.5.8
PyPDF2==3.0.1
python-docx==1.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
werkzeug==3.0.1